refactor(training-link): add src and types folders
