--- packages/training-link/TrainingLink.d.ts
@@ -1,8 +1,0 @@
-export interface TrainingLinkProps {
-    link: string;
-    name: string;
-}
-
-declare const TrainingLink: React.FunctionComponent<TrainingLinkProps>;
-
-export default TrainingLink;
