+++ packages/training-link/src/TrainingLink.js
