+++ packages/training-link/types/TrainingLink.d.ts
@@ -1,0 +1,8 @@
+export interface TrainingLinkProps {
+  link: string;
+  name: string;
+}
+
+declare const TrainingLink: React.FC<TrainingLinkProps>;
+
+export default TrainingLink;
